Uploaded Code and images v1.0
This version of the code contains the ability to generate teams by checking who is playing in the 10 man. As of right now it randomly selects team and map banning is automatic, still trying to convert 3 of my codes into one functional code. Halftime is not working in this commit